refactor: splitting user tests
