+++ src/api/schema/user/user.test.ts
@@ -5,8 +5,6 @@ import { expect } from 'chai';
 import { UserInput } from '@api/schema/user/user.input';
 import { UserEntity } from '@data/entity/user.entity';
 import { Authenticator } from '@api/server/authenticator';
-import { UserSeed } from '@data/seed/user.seed';
-import { UserType } from './user.type';
 
 const userQuery = `
 query user ($id: String!) {
@@ -18,38 +16,6 @@ query user ($id: String!) {
   }
 }`;
 
-const usersQuery = `
-query users ($limit: Int) {
-  users(limit : $limit) {
-    id
-    name
-    email
-    birthDate
-  }
-}`;
-
-const createUserMutation = `
-mutation createUser($data: UserInput!) {
-  createUser(data: $data) {
-    id
-    name
-    email
-    birthDate
-  }
-}`;
-
-const getTestToken = async (): Promise<string> => {
-  const user = UserEntity.create({
-    name: 'Luke Skywalker',
-    email: 'skylwalker.top@gmail.com',
-    password: 'a1ÊÇ7ma2',
-    birthDate: new Date(),
-  });
-  await user.save();
-
-  return Authenticator.getJWT({ id: user.id });
-};
-
 describe('GraphQL: User - query user', () => {
   it('should successfully return a user', async () => {
     const user = UserEntity.create({
@@ -60,7 +26,7 @@ describe('GraphQL: User - query user', () => {
     });
     await user.save();
 
-    const token = await getTestToken();
+    const token = await Authenticator.getTestToken();
     const res = await Request(userQuery, { id: user.id }, token);
 
     expect(res.body).to.not.own.property('errors');
@@ -73,7 +39,7 @@ describe('GraphQL: User - query user', () => {
   });
 
   it('should trigger unknown user error', async () => {
-    const token = await getTestToken();
+    const token = await Authenticator.getTestToken();
     const res = await Request(userQuery, { id: '00000000-0000-0000-0000-000000000000' }, token);
 
     expect(res.body.data).to.be.null;
@@ -146,239 +112,5 @@ describe('GraphQL: User - query user', () => {
       message: 'Usuário não autorizado',
       details: 'Token inválido',
     });
-  });
-});
-
-describe('GraphQL: User - query users', function () {
-  this.timeout(5000);
-
-  it('should successfully return 10 users without a defined limit', async () => {
-    await UserSeed(10);
-
-    const res = await Request(usersQuery);
-
-    expect(res.body).to.not.own.property('errors');
-    expect(res.body.data.users).to.have.lengthOf(10);
-
-    const reorderedUsers = res.body.data.users;
-    reorderedUsers.sort((userA: UserType, userB: UserType) => {
-      var nameA = userA.name.toUpperCase();
-      var nameB = userB.name.toUpperCase();
-      return nameA < nameB ? -1 : nameA > nameB ? 1 : 0;
-    });
-
-    for (let i = 0; i < res.body.data.users.length; i++) {
-      expect(res.body.data.users[i]).to.have.all.keys('id', 'name', 'email', 'birthDate');
-      expect(res.body.data.users[i]).to.be.eq(reorderedUsers[i]);
-    }
-  });
-
-  it('should successfully return 5 users with a defined limit', async () => {
-    await UserSeed(10);
-
-    const res = await Request(usersQuery, { limit: 5 });
-
-    expect(res.body).to.not.own.property('errors');
-    expect(res.body.data.users).to.have.lengthOf(5);
-
-    const reorderedUsers = res.body.data.users;
-    reorderedUsers.sort((userA: UserType, userB: UserType) => {
-      var nameA = userA.name.toUpperCase();
-      var nameB = userB.name.toUpperCase();
-      return nameA < nameB ? -1 : nameA > nameB ? 1 : 0;
-    });
-
-    for (let i = 0; i < res.body.data.users.length; i++) {
-      expect(res.body.data.users[i]).to.have.all.keys('id', 'name', 'email', 'birthDate');
-      expect(res.body.data.users[i]).to.be.eq(reorderedUsers[i]);
-    }
-  });
-
-  it('it should successfully return 10 users with an overly defined limit', async () => {
-    await UserSeed(10);
-
-    const res = await Request(usersQuery, { limit: 100 });
-
-    expect(res.body).to.not.own.property('errors');
-    expect(res.body.data.users).to.have.lengthOf(10);
-
-    const reorderedUsers = res.body.data.users;
-    reorderedUsers.sort((userA: UserType, userB: UserType) => {
-      var nameA = userA.name.toUpperCase();
-      var nameB = userB.name.toUpperCase();
-      return nameA < nameB ? -1 : nameA > nameB ? 1 : 0;
-    });
-
-    for (let i = 0; i < res.body.data.users.length; i++) {
-      expect(res.body.data.users[i]).to.have.all.keys('id', 'name', 'email', 'birthDate');
-      expect(res.body.data.users[i]).to.be.eq(reorderedUsers[i]);
-    }
-  });
-
-  it('it should trigger non-positive error', async () => {
-    const res = await Request(usersQuery, { limit: -10 });
-
-    expect(res.body.data).to.be.null;
-    expect(res.body.errors).to.deep.include({
-      code: 400,
-      message: 'O limite não pode ser negativo',
-    });
-  });
-
-  it('it should trigger non-integer error', async () => {
-    const res = await Request(usersQuery, { limit: 4.2 }, undefined, 400);
-    expect(res.body.errors[0].message).to.includes('Int cannot represent non-integer value');
-  });
-});
-
-describe('GraphQL: User - mutation createUser', () => {
-  it('should successfully create user', async () => {
-    const input: UserInput = {
-      name: 'Padmé Amidala',
-      email: 'padmeia@yahoo.com',
-      password: 'padead123',
-      birthDate: new Date(),
-    };
-
-    const token = await getTestToken();
-    const res = await Request(createUserMutation, { data: input }, token);
-
-    expect(res.body).to.not.own.property('errors');
-    expect(res.body.data.createUser).to.have.property('id');
-    expect(res.body.data.createUser).to.include({
-      name: input.name,
-      email: input.email,
-      birthDate: input.birthDate.toISOString(),
-    });
-
-    const user = await UserEntity.findOneOrFail(res.body.data.createUser.id);
-    expect(user).to.not.be.undefined;
-    expect(await bcrypt.compare(input.password, user.password)).to.be.true;
-    expect(user).to.deep.include({
-      id: res.body.data.createUser.id,
-      name: input.name,
-      email: input.email,
-      birthDate: input.birthDate,
-    });
-  });
-
-  it('should trigger token not sent error', async () => {
-    const input: UserInput = {
-      name: 'Padmé Amidala',
-      email: 'wrong email',
-      password: 'padead123',
-      birthDate: new Date(),
-    };
-
-    const res = await Request(createUserMutation, { data: input });
-
-    expect(res.body.data).to.be.null;
-    expect(res.body.errors).to.deep.include({
-      code: 401,
-      message: 'Usuário não autorizado',
-      details: 'Token não enviado',
-    });
-  });
-
-  it('should trigger expired token error', async () => {
-    const input: UserInput = {
-      name: 'Padmé Amidala',
-      email: 'wrong email',
-      password: 'padead123',
-      birthDate: new Date(),
-    };
-
-    const res = await Request(
-      createUserMutation,
-      { data: input },
-      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImQ2ODI4MmU4LThlNTEtNGNiZi04MzBlLTg1NGZhNzFkOWJhYiIsImlhdCI6MTYxNTIyMjk4NCwiZXhwIjoxNjE1MjIyOTg2fQ._6-JMIVvkJVVhr8ic3qzTDHSpAAvibL54xWLVW1u-TU',
-    );
-
-    expect(res.body.data).to.be.null;
-    expect(res.body.errors).to.deep.include({
-      code: 401,
-      message: 'Usuário não autorizado',
-      details: 'Token expirado',
-    });
-  });
-
-  it('should trigger invalid token error', async () => {
-    const input: UserInput = {
-      name: 'Padmé Amidala',
-      email: 'wrong email',
-      password: 'padead123',
-      birthDate: new Date(),
-    };
-
-    const res = await Request(
-      createUserMutation,
-      { data: input },
-      'eyJpZCI6ImQ2ODI4MmU4LThlNTEtNGNiZi04MzBlLTg1NGZhNzFkOWJhYiIsImlhdCI6MTYxNTIyMjk4NCwiZXhwIjoxNjE1MjIyOTg2fQ._6-JMIVvkJVVhr8ic3qzTDHSpAAvibL54xWLVW1u-TU',
-    );
-
-    expect(res.body.data).to.be.null;
-    expect(res.body.errors).to.deep.include({
-      code: 401,
-      message: 'Usuário não autorizado',
-      details: 'Token inválido',
-    });
-  });
-
-  it('should trigger duplicate email error', async () => {
-    const input: UserInput = {
-      name: 'Luke Skywalker',
-      email: 'skylwalker.top@gmail.com',
-      password: 'a1ÊÇ7ma2',
-      birthDate: new Date(),
-    };
-
-    const token = await getTestToken();
-    const res = await Request(createUserMutation, { data: input }, token);
-
-    expect(res.body.data).to.be.null;
-    expect(res.body.errors).to.deep.include({ code: 400, message: 'Email já cadastrado' });
-  });
-
-  it('should trigger email validation error', async () => {
-    const input: UserInput = {
-      name: 'Padmé Amidala',
-      email: 'wrong email',
-      password: 'padead123',
-      birthDate: new Date(),
-    };
-
-    const token = await getTestToken();
-    const res = await Request(createUserMutation, { data: input }, token);
-
-    expect(res.body.data).to.be.null;
-
-    const errorMessages = res.body.errors.map((error: { message: string }) => error.message);
-    expect(errorMessages).to.include('Argumentos inválidos');
-    const errorIndex = errorMessages.indexOf('Argumentos inválidos');
-
-    expect(res.body.errors[errorIndex]).to.own.property('details');
-    expect(res.body.errors[errorIndex].details).to.include('O email precisa ser um endereço de e-mail válido');
-  });
-
-  it('should trigger password validation error', async () => {
-    const input: UserInput = {
-      name: 'Padmé Amidala',
-      email: 'padmeia@yahoo.com',
-      password: 'aaa',
-      birthDate: new Date(),
-    };
-
-    const token = await getTestToken();
-    const res = await Request(createUserMutation, { data: input }, token);
-
-    expect(res.body.data).to.be.null;
-
-    const errorMessages = res.body.errors.map((error: { message: string }) => error.message);
-    expect(errorMessages).to.include('Argumentos inválidos');
-    const errorIndex = errorMessages.indexOf('Argumentos inválidos');
-
-    expect(res.body.errors[errorIndex]).to.own.property('details');
-    expect(res.body.errors[errorIndex].details).to.include('A senha precisa ter pelo menos 7 caracteres');
-    expect(res.body.errors[errorIndex].details).to.include('A senha precisa ter pelo uma letra e um número');
   });
 });
